Simplifying using bot
